+++ src/audio_metadata/formats/wav.py
@@ -4,7 +4,7 @@ import struct
 
 from attr import attrib, attrs
 
-from .id3 import ID3v2, ID3v2Frames
+from .id3v2 import ID3v2, ID3v2Frames
 from .models import Format, StreamInfo
 from ..exceptions import InvalidFrame, InvalidHeader